[fix] GET 할 때 id도 조회

+++ api/src/main/java/org/coteis/domain/Article.java
@@ -16,7 +16,7 @@ public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "article_no", updatable = false)
-    private Integer articleNo;
+    private Long articleNo;
 
     @Column(name = "title", nullable = false)   // 제목
     private String title;
